safari-600.7-branch/LayoutTests/inspector/debugger/probe-manager-add-remove-actions-expected.txt

Testing that the probe manager properly handles addition and removal of related probes.

Breakpoint autocontinue state changed: true
Probe set was added. New count: 1
Probe with identifier 1 was added to probe set.
Probe set's probe count: 1
Breakpoint actions changed. New count: 1
Probe with identifier 2 was added to probe set.
Probe set's probe count: 2
Breakpoint actions changed. New count: 2
Breakpoint resolved state changed: true
Breakpoint resolved state changed: true
Breakpoint resolved state changed: false
Breakpoint was added.
PASS: Breakpoint should be resolved.
Breakpoint resolved state changed: true
Probe with identifier 1 added sample: [object Object]
Hit test checkpoint event #0 with type: probe-object-sample-added
Probe with identifier 2 added sample: [object Object]
Hit test checkpoint event #1 with type: probe-object-sample-added
